Add function to Lox
When a function is declared, a function object with parameters, body and current environment (closure) will be generated and put in current environment as value with its name as key. When a function is called, parameters are evaluated and added to the closure with formal corresponding parameters. Then the body is executed (as body statement) with closure as new environment of interpreter.

Others.
- native function
- return statement using Java exception in Return.java
- function object structure in LoxFunction.java