Add local variable for clox
local variables are stored in compiler, which have the same index as
stack in vm. When emiting op code, the index directly emits as
following. So, vm just use the poisition in stack to access corresponding local variable. That is, vm does not has local variable, instead of fetching value of stack.